+++ ghost/core/package.json
@@ -112,7 +112,7 @@
     "@tryghost/members-ssr": "0.0.0",
     "@tryghost/members-stripe-service": "0.0.0",
     "@tryghost/metrics": "1.0.20",
-    "@tryghost/milestone-emails": "0.0.0",
+    "@tryghost/milestones": "0.0.0",
     "@tryghost/minifier": "0.0.0",
     "@tryghost/mw-api-version-mismatch": "0.0.0",
     "@tryghost/mw-cache-control": "0.0.0",
