+++ ghost/core/test/e2e-server/services/milestones.test.js
@@ -5,7 +5,7 @@ const sinon = require('sinon');
 const models = require('../../../core/server/models');
 const moment = require('moment');
 
-const milestoneEmailsService = require('../../../core/server/services/milestone-emails');
+const milestonesService = require('../../../core/server/services/milestones');
 
 let agent;
 let counter = 0;
@@ -137,7 +137,7 @@ async function createFreeMembers(amount, amountImported = 0) {
     await Promise.all(members);
 }
 
-describe('Milestone Emails Service', function () {
+describe('Milestones Service', function () {
     const milestonesConfig = {
         arr: [{currency: 'usd', values: [100, 150]}],
         members: [10, 20, 30]
@@ -153,7 +153,7 @@ describe('Milestone Emails Service', function () {
         sinon.createSandbox();
         // TODO: stub out stripe mode
         // stripeModeStub = sinon.stub().returns(true);
-        // milestoneEmailsService.__set__('getStripeLiveEnabled', stripeModeStub);
+        // milestonesService.__set__('getStripeLiveEnabled', stripeModeStub);
         configUtils.set('milestones', milestonesConfig);
         mockManager.mockLabsEnabled('milestoneEmails');
     });
@@ -165,29 +165,29 @@ describe('Milestone Emails Service', function () {
     });
 
     it('Inits milestone service', async function () {
-        await milestoneEmailsService.init();
+        await milestonesService.init();
 
-        assert.ok(milestoneEmailsService.api);
+        assert.ok(milestonesService.api);
     });
 
     it('Runs ARR and Members milestone jobs', async function () {
         mockManager.mockSetting('stripe_connect_publishable_key', 'pk_live_89843uihsidfh98832uo8ri');
 
         // No ARR and no members
-        const firstRun = await milestoneEmailsService.initAndRun();
+        const firstRun = await milestonesService.initAndRun();
         assert(firstRun.members === undefined);
         assert(firstRun.arr === undefined);
 
         await createFreeMembers(7);
         await createMemberWithSubscription('year', 5000, 'usd', '2000-01-10');
         await createMemberWithSubscription('month', 100, 'usd', '2000-01-10');
-        const secondRun = await milestoneEmailsService.initAndRun();
+        const secondRun = await milestonesService.initAndRun();
         assert(secondRun.members === undefined);
         assert(secondRun.arr === undefined);
 
         // Reached the first milestone for members
         await createFreeMembers(1);
-        const thirdRun = await milestoneEmailsService.initAndRun();
+        const thirdRun = await milestonesService.initAndRun();
         assert(thirdRun.members.value === 10);
         assert(thirdRun.members.emailSentAt !== undefined);
         assert(thirdRun.arr === undefined);
@@ -197,7 +197,7 @@ describe('Milestone Emails Service', function () {
         // will be created
         await createMemberWithSubscription('month', 500, 'usd', '2000-01-10');
         await createMemberWithSubscription('month', 500, 'eur', '2000-01-10');
-        const fourthRun = await milestoneEmailsService.initAndRun();
+        const fourthRun = await milestonesService.initAndRun();
         assert(fourthRun.members === undefined);
         assert(fourthRun.arr.value === 100);
         assert(fourthRun.arr.emailSentAt !== undefined);
@@ -209,7 +209,7 @@ describe('Milestone Emails Service', function () {
 
         await createFreeMembers(10, 1);
         await createMemberWithSubscription('month', 1000, 'usd', '2023-01-10');
-        const result = await milestoneEmailsService.initAndRun();
+        const result = await milestonesService.initAndRun();
 
         assert(result.members.value === 20);
         assert(result.members.emailSentAt === null);
@@ -220,7 +220,7 @@ describe('Milestone Emails Service', function () {
     it('Does not run when milestoneEmails labs flag is not set', async function () {
         mockManager.mockLabsDisabled('milestoneEmails');
 
-        const result = await milestoneEmailsService.initAndRun();
+        const result = await milestonesService.initAndRun();
         assert(result === undefined);
     });
 
@@ -229,7 +229,7 @@ describe('Milestone Emails Service', function () {
         mockManager.mockSetting('stripe_connect_publishable_key', 'pk_test_89843uihsidfh98832uo8ri');
         await createFreeMembers(10);
 
-        const result = await milestoneEmailsService.initAndRun();
+        const result = await milestonesService.initAndRun();
         assert(result.members.value === 30);
         assert(result.members.emailSentAt !== undefined);
         assert(result.arr === undefined);
