--- ghost/core/test/unit/server/services/milestone-emails/index.test.js
@@ -1,13 +1,0 @@
-const assert = require('assert');
-
-describe('Milestone Emails Service', function () {
-    let milestoneEmails;
-
-    describe('Milestone Emails Service', function () {
-        it('Provides expected public API', async function () {
-            milestoneEmails = require('../../../../../core/server/services/milestone-emails');
-
-            assert.ok(milestoneEmails.initAndRun);
-        });
-    });
-});
