+++ ghost/milestones/lib/Milestone.js
