+++ ghost/milestones/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@tryghost/milestone-emails",
+  "name": "@tryghost/milestones",
   "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/milestone-emails",
+  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/milestones",
   "author": "Ghost Foundation",
   "private": true,
   "main": "index.js",
